services/obituary-intelligence-engine/pyproject.toml
[tool.poetry]
name = "obituary-intelligence-engine"
version = "0.1.0"
description = "Iowa obituary intelligence service for lli-saas"
authors = ["nice-and-precise"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
beautifulsoup4 = "^4.12.3"
feedparser = "^6.0.11"
google-genai = "^1.8.0"
anthropic = "^0.49.0"
lxml = "^5.3.1"
nameparser = "^1.1.3"
rapidfuzz = "^3.11.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
